CTkinterUI/requirements.txt
customtkinter
requests
pytorch
torchvision
torchaudio
transformers
google-generativeai
anthropic
openai
groq